Correct the filename in device storage (mimetype null)
This correction allow the developer define the correct name of the file (with the extension) to be write in the device storage when the URL is a file dinamically generated. In my case my URL was:
http://domain.server/dir1/dir2/dir3/generate_file.php?key=1234&param1=abc&param2=xyz
Without my correction i tryed this:
http://domain.server/dir1/dir2/dir3/generate_file.php?key=1234&param1=abc&param2=xyz&filename.txt
But the mimetype becames null, now with this little modification, the mimetype became correct.